// syncro/apps/web/src/lib/api/generated/model/createMachineResponsibilityRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { CreateMachineResponsibilityRequestLevel } from './createMachineResponsibilityRequestLevel';

export interface CreateMachineResponsibilityRequest {
  machineId: string;
  userId: string;
  level: CreateMachineResponsibilityRequestLevel;
}